地形に凹角(concave)形状を追加しボクセルを2byte(ushort)化
- 形状18種に拡張(concave×4追加)し4bitに収まらないためボクセルを2byte化
  ("LWVT" v2・RLEは(ushort値LE + byte count)の3バイトペア・リリース前のため後方互換なし)
- 凹角のメッシュ生成/コライダー(L字2box×3段)/近傍カリング/タイプ変更サイクル/配置条件を実装
- 坂の低い側が反対向きの坂でも開き扱い(隣接2坂でV字の溝)
- 上面/下面ピック(TryPickEditCell・編集開始時スナップショット固定)とマウス右ボタンパンを追加
- 特殊オブジェクト重複判定 SpecialObjectOverlap + WorldPublishValidator.SpawnPortalOverlap を追加
- 関連docs(world-creation/screens-and-modes/development-plan)を更新

+++ LowPolyWorldUnity/Assets/Scripts/World/Terrain/TerrainEditScenePreview.cs
@@ -35,7 +35,7 @@ public class TerrainEditScenePreview : MonoBehaviour
 
         // サンプル地形: 全面の地面（高さ 0）
         var store = new TerrainVoxelStore();
-        byte ground = TerrainVoxel.Encode(TerrainShape.Cube, 0);
+        ushort ground = TerrainVoxel.Encode(TerrainShape.Cube, 0);
         for (int x = 0; x < TerrainVoxelStore.SizeX; x++)
             for (int z = 0; z < TerrainVoxelStore.SizeZ; z++)
                 store.SetVoxel(x, 0, z, ground);
