+++ LowPolyWorldUnity/Assets/Tests/EditMode/TerrainVoxelStoreTests.cs
@@ -20,8 +20,8 @@ public class TerrainVoxelStoreTests
     [Test]
     public void SetGetVoxel_AcrossChunkBoundaries()
     {
-        byte a = TerrainVoxel.Encode(TerrainShape.Cube, 1);
-        byte b = TerrainVoxel.Encode(TerrainShape.RampW, 2);
+        ushort a = TerrainVoxel.Encode(TerrainShape.Cube, 1);
+        ushort b = TerrainVoxel.Encode(TerrainShape.RampW, 2);
 
         _store.SetVoxel(15, 0, 0, a); // チャンク (0,0,0) の端
         _store.SetVoxel(16, 0, 0, b); // チャンク (1,0,0) の先頭
@@ -63,7 +63,7 @@ public class TerrainVoxelStoreTests
     [Test]
     public void EnumerateNonEmptyChunks_SkipsEmptiedChunks()
     {
-        byte v = TerrainVoxel.Encode(TerrainShape.Cube, 0);
+        ushort v = TerrainVoxel.Encode(TerrainShape.Cube, 0);
         _store.SetVoxel(0, 0, 0, v);
         _store.SetVoxel(32, 0, 0, v);
         _store.SetVoxel(32, 0, 0, TerrainVoxel.Empty); // チャンク (2,0,0) を空に戻す
